feat: move wysiwyg dom helper to utils
